LBaaS v2 Shared Functions
Moves all shared functions into a dedicated file in hopes of
simplifying maintenance of the LBaaSv2 resources.
