+++ openstack/resource_openstack_lb_loadbalancer_v2.go
@@ -286,42 +286,3 @@ func resourceLoadBalancerV2SecurityGroups(networkingClient *gophercloud.ServiceC
 
 	return nil
 }
-
-func resourceLBV2LoadBalancerRefreshFunc(networkingClient *gophercloud.ServiceClient, id string) resource.StateRefreshFunc {
-	return func() (interface{}, string, error) {
-		lb, err := loadbalancers.Get(networkingClient, id).Extract()
-		if err != nil {
-			return nil, "", err
-		}
-
-		return lb, lb.ProvisioningStatus, nil
-	}
-}
-
-func waitForLBV2LoadBalancer(networkingClient *gophercloud.ServiceClient, id string, target string, pending []string, timeout time.Duration) error {
-	log.Printf("[DEBUG] Waiting for loadbalancer %s to become %s.", id, target)
-
-	stateConf := &resource.StateChangeConf{
-		Target:     []string{target},
-		Pending:    pending,
-		Refresh:    resourceLBV2LoadBalancerRefreshFunc(networkingClient, id),
-		Timeout:    timeout,
-		Delay:      5 * time.Second,
-		MinTimeout: 1 * time.Second,
-	}
-
-	_, err := stateConf.WaitForState()
-	if err != nil {
-		if _, ok := err.(gophercloud.ErrDefault404); ok {
-			switch target {
-			case "DELETED":
-				return nil
-			default:
-				return fmt.Errorf("Error: loadbalancer %s not found: %s", id, err)
-			}
-		}
-		return fmt.Errorf("Error waiting for loadbalancer %s to become %s: %s", id, target, err)
-	}
-
-	return nil
-}
